refactor: 컨트롤러 파일들을 용도에 맞는 위치로 변경
